+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.Application/MarketParticipants/Handlers/Mappers/MarketParticipantIntegrationEventMapper.cs
@@ -48,6 +48,7 @@ namespace GreenEnergyHub.Charges.Application.MarketParticipants.Handlers.Mappers
             return new MarketParticipantCreatedCommand(
                 actorCreatedIntegrationEvent.ActorId,
                 actorCreatedIntegrationEvent.ActorNumber,
+                actorCreatedIntegrationEvent.Name,
                 roles,
                 status,
                 actorCreatedIntegrationEvent.ActorMarketRoles
@@ -90,6 +91,14 @@ namespace GreenEnergyHub.Charges.Application.MarketParticipants.Handlers.Mappers
         {
             return new GridAreaRemovedFromMarketParticipantCommand(
                 gridUpdatedIntegrationEvent.GridAreaId);
+        }
+
+        public static MarketParticipantNameChangedCommand Map(
+            ActorNameChangedIntegrationEvent actorNameChangedIntegrationEvent)
+        {
+            return new MarketParticipantNameChangedCommand(
+                actorNameChangedIntegrationEvent.ActorId,
+                actorNameChangedIntegrationEvent.Name);
         }
     }
 }