+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.FunctionHost/Common/EnvironmentSettingNames.cs
@@ -67,9 +67,10 @@ namespace GreenEnergyHub.Charges.FunctionHost.Common
         public const string MarketParticipantChangedSubscriptionName = "MARKET_PARTICIPANT_CHANGED_SUBSCRIPTION_NAME";
         public const string MarketParticipantCreatedSubscriptionName = "MARKET_PARTICIPANT_CREATED_SUBSCRIPTION_NAME";
         public const string MarketParticipantStatusChangedSubscriptionName = "MARKET_PARTICIPANT_STATUS_CHANGED_SUBSCRIPTION_NAME";
-        public const string MarketParticipantExternalActorIdChangedSubscriptionName = "MARKET_PARTICIPANT_STATUS_EXTERNAL_ACTOR_ID_CHANGED_SUBSCRIPTION_NAME";
         public const string GridAreaAddedToActorSubscriptionName = "GRID_AREA_ADDED_TO_ACTOR_SUBSCRIPTION_NAME";
         public const string GridAreaRemovedFromActorSubscriptionName = "GRID_AREA_REMOVED_FROM_ACTOR_SUBSCRIPTION_NAME";
+        public const string MarketParticipantExternalActorIdChangedSubscriptionName = "MARKET_PARTICIPANT_EXTERNAL_ACTOR_ID_CHANGED_SUBSCRIPTION_NAME";
+        public const string MarketParticipantNameChangedSubscriptionName = "MARKET_PARTICIPANT_NAME_CHANGED_SUBSCRIPTION_NAME";
 
         // Domain event
         [DomainEventSetting]