+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.IntegrationTests/IntegrationTests/WebApi/QueryServices/ChargesQueryServiceTests.cs
@@ -18,7 +18,6 @@ using System.Linq;
 using System.Threading.Tasks;
 using Energinet.DataHub.Charges.Contracts.Charge;
 using FluentAssertions;
-using GreenEnergyHub.Charges.Infrastructure.Persistence;
 using GreenEnergyHub.Charges.IntegrationTest.Core.Fixtures.Database;
 using GreenEnergyHub.Charges.QueryApi;
 using GreenEnergyHub.Charges.QueryApi.Model;
@@ -26,9 +25,7 @@ using GreenEnergyHub.Charges.QueryApi.QueryServices;
 using GreenEnergyHub.Charges.TestCore;
 using GreenEnergyHub.Charges.TestCore.Builders.Command;
 using GreenEnergyHub.Charges.TestCore.Builders.Query;
-using GreenEnergyHub.Charges.TestCore.Builders.Testables;
 using GreenEnergyHub.TestHelpers;
-using Microsoft.EntityFrameworkCore;
 using NodaTime.Extensions;
 using Xunit;
 using Xunit.Categories;
@@ -41,7 +38,6 @@ namespace GreenEnergyHub.Charges.IntegrationTests.IntegrationTests.WebApi.QueryS
     [IntegrationTest]
     public class ChargesQueryServiceTests : IClassFixture<ChargesDatabaseFixture>
     {
-        private const string MarketParticipantOwnerId = "MarketParticipantId";
         private readonly ChargesDatabaseManager _databaseManager;
 
         public ChargesQueryServiceTests(ChargesDatabaseFixture fixture)
@@ -54,7 +50,7 @@ namespace GreenEnergyHub.Charges.IntegrationTests.IntegrationTests.WebApi.QueryS
         {
             // Arrange
             await using var chargesDatabaseWriteContext = _databaseManager.CreateDbContext();
-            var marketParticipantId = await GetOrAddMarketParticipantAsync(chargesDatabaseWriteContext, MarketParticipantOwnerId);
+            var marketParticipantId = await QueryServiceAutoMoqDataFixer.GetOrAddMarketParticipantAsync(chargesDatabaseWriteContext);
 
             var charge = GetValidCharge(marketParticipantId);
 
@@ -62,7 +58,7 @@ namespace GreenEnergyHub.Charges.IntegrationTests.IntegrationTests.WebApi.QueryS
             await chargesDatabaseWriteContext.SaveChangesAsync();
 
             await using var chargesDatabaseQueryContext = _databaseManager.CreateDbQueryContext();
-            var searchCriteria = new SearchCriteriaDtoBuilder().Build();
+            var searchCriteria = new ChargeSearchCriteriaV1DtoBuilder().Build();
 
             var expected = GetActiveCharges(chargesDatabaseQueryContext).Count();
             var sut = GetSut(chargesDatabaseQueryContext);
@@ -80,7 +76,7 @@ namespace GreenEnergyHub.Charges.IntegrationTests.IntegrationTests.WebApi.QueryS
         {
             // Arrange
             await using var chargesDatabaseWriteContext = _databaseManager.CreateDbContext();
-            var marketParticipantId = await GetOrAddMarketParticipantAsync(chargesDatabaseWriteContext, MarketParticipantOwnerId);
+            var marketParticipantId = await QueryServiceAutoMoqDataFixer.GetOrAddMarketParticipantAsync(chargesDatabaseWriteContext);
 
             var charge = GetValidCharge(marketParticipantId);
 
@@ -89,7 +85,7 @@ namespace GreenEnergyHub.Charges.IntegrationTests.IntegrationTests.WebApi.QueryS
             var expected = charge.SenderProvidedChargeId;
 
             await using var chargesDatabaseQueryContext = _databaseManager.CreateDbQueryContext();
-            var searchCriteria = new SearchCriteriaDtoBuilder()
+            var searchCriteria = new ChargeSearchCriteriaV1DtoBuilder()
                 .WithChargeIdOrName(expected)
                 .Build();
 
@@ -113,7 +109,7 @@ namespace GreenEnergyHub.Charges.IntegrationTests.IntegrationTests.WebApi.QueryS
         {
             // Arrange
             await using var chargesDatabaseWriteContext = _databaseManager.CreateDbContext();
-            var marketParticipantId = await GetOrAddMarketParticipantAsync(chargesDatabaseWriteContext, MarketParticipantOwnerId);
+            var marketParticipantId = await QueryServiceAutoMoqDataFixer.GetOrAddMarketParticipantAsync(chargesDatabaseWriteContext);
 
             var todayAtMidnightUtc = DateTime.Now.Date.ToUniversalTime().ToInstant();
 
@@ -125,7 +121,7 @@ namespace GreenEnergyHub.Charges.IntegrationTests.IntegrationTests.WebApi.QueryS
             && c.Periods.Any(p => p.StartDateTime <= todayAtMidnightUtc));
 
             await using var chargesDatabaseQueryContext = _databaseManager.CreateDbQueryContext();
-            var searchCriteria = new SearchCriteriaDtoBuilder()
+            var searchCriteria = new ChargeSearchCriteriaV1DtoBuilder()
                 .WithChargeType(expectedChargeType)
                 .Build();
 
@@ -144,7 +140,7 @@ namespace GreenEnergyHub.Charges.IntegrationTests.IntegrationTests.WebApi.QueryS
         {
             // Arrange
             await using var chargesDatabaseWriteContext = _databaseManager.CreateDbContext();
-            var marketParticipantId = await GetOrAddMarketParticipantAsync(chargesDatabaseWriteContext, MarketParticipantOwnerId);
+            var marketParticipantId = await QueryServiceAutoMoqDataFixer.GetOrAddMarketParticipantAsync(chargesDatabaseWriteContext);
 
             var charge = GetValidCharge(marketParticipantId);
 
@@ -154,7 +150,7 @@ namespace GreenEnergyHub.Charges.IntegrationTests.IntegrationTests.WebApi.QueryS
             await using var chargesDatabaseQueryContext = _databaseManager.CreateDbQueryContext();
             var expected = GetActiveCharges(chargesDatabaseQueryContext).Count();
 
-            var searchCriteria = new SearchCriteriaDtoBuilder()
+            var searchCriteria = new ChargeSearchCriteriaV1DtoBuilder()
                 .WithChargeTypes(null!)
                 .Build();
 
@@ -173,7 +169,7 @@ namespace GreenEnergyHub.Charges.IntegrationTests.IntegrationTests.WebApi.QueryS
         {
             // Arrange
             await using var chargesDatabaseWriteContext = _databaseManager.CreateDbContext();
-            var marketParticipantId = await GetOrAddMarketParticipantAsync(chargesDatabaseWriteContext, MarketParticipantOwnerId);
+            var marketParticipantId = await QueryServiceAutoMoqDataFixer.GetOrAddMarketParticipantAsync(chargesDatabaseWriteContext);
 
             var charge = GetValidCharge(marketParticipantId);
 
@@ -183,7 +179,7 @@ namespace GreenEnergyHub.Charges.IntegrationTests.IntegrationTests.WebApi.QueryS
                 c.Type == Domain.Charges.ChargeType.Subscription || c.Type == Domain.Charges.ChargeType.Fee);
 
             await using var chargesDatabaseQueryContext = _databaseManager.CreateDbQueryContext();
-            var searchCriteria = new SearchCriteriaDtoBuilder()
+            var searchCriteria = new ChargeSearchCriteriaV1DtoBuilder()
                 .WithChargeTypes(new List<ChargeType> { ChargeType.D01, ChargeType.D02 })
                 .Build();
 
@@ -202,7 +198,7 @@ namespace GreenEnergyHub.Charges.IntegrationTests.IntegrationTests.WebApi.QueryS
         {
             // Arrange
             await using var chargesDatabaseWriteContext = _databaseManager.CreateDbContext();
-            var marketParticipantId = await GetOrAddMarketParticipantAsync(chargesDatabaseWriteContext, MarketParticipantOwnerId);
+            var marketParticipantId = await QueryServiceAutoMoqDataFixer.GetOrAddMarketParticipantAsync(chargesDatabaseWriteContext);
 
             var charge = GetValidCharge(marketParticipantId);
 
@@ -211,7 +207,7 @@ namespace GreenEnergyHub.Charges.IntegrationTests.IntegrationTests.WebApi.QueryS
             var expected = charge.Periods.First().Name;
 
             await using var chargesDatabaseQueryContext = _databaseManager.CreateDbQueryContext();
-            var searchCriteria = new SearchCriteriaDtoBuilder()
+            var searchCriteria = new ChargeSearchCriteriaV1DtoBuilder()
                 .WithChargeIdOrName(expected)
                 .Build();
 
@@ -230,7 +226,7 @@ namespace GreenEnergyHub.Charges.IntegrationTests.IntegrationTests.WebApi.QueryS
         {
             // Arrange
             await using var chargesDatabaseWriteContext = _databaseManager.CreateDbContext();
-            var marketParticipantId = await GetOrAddMarketParticipantAsync(chargesDatabaseWriteContext, MarketParticipantOwnerId);
+            var marketParticipantId = await QueryServiceAutoMoqDataFixer.GetOrAddMarketParticipantAsync(chargesDatabaseWriteContext);
 
             var expected = "test_with_many_periods_get_one";
             var charge = GetValidChargeWithMultiplePeriods(marketParticipantId, expected);
@@ -239,7 +235,7 @@ namespace GreenEnergyHub.Charges.IntegrationTests.IntegrationTests.WebApi.QueryS
             await chargesDatabaseWriteContext.SaveChangesAsync();
 
             await using var chargesDatabaseQueryContext = _databaseManager.CreateDbQueryContext();
-            var searchCriteria = new SearchCriteriaDtoBuilder()
+            var searchCriteria = new ChargeSearchCriteriaV1DtoBuilder()
                 .WithChargeIdOrName(expected)
                 .Build();
 
@@ -258,7 +254,7 @@ namespace GreenEnergyHub.Charges.IntegrationTests.IntegrationTests.WebApi.QueryS
         {
             // Arrange
             await using var chargesDatabaseWriteContext = _databaseManager.CreateDbContext();
-            var marketParticipantId = await GetOrAddMarketParticipantAsync(chargesDatabaseWriteContext, MarketParticipantOwnerId);
+            var marketParticipantId = await QueryServiceAutoMoqDataFixer.GetOrAddMarketParticipantAsync(chargesDatabaseWriteContext);
 
             var charge = GetValidCharge(marketParticipantId);
 
@@ -268,7 +264,7 @@ namespace GreenEnergyHub.Charges.IntegrationTests.IntegrationTests.WebApi.QueryS
             var expected = charge.SenderProvidedChargeId;
 
             await using var chargesDatabaseQueryContext = _databaseManager.CreateDbQueryContext();
-            var searchCriteria = new SearchCriteriaDtoBuilder()
+            var searchCriteria = new ChargeSearchCriteriaV1DtoBuilder()
                 .WithChargeIdOrName(searchName)
                 .Build();
 
@@ -287,7 +283,7 @@ namespace GreenEnergyHub.Charges.IntegrationTests.IntegrationTests.WebApi.QueryS
         {
             // Arrange
             await using var chargesDatabaseWriteContext = _databaseManager.CreateDbContext();
-            var marketParticipantId = await GetOrAddMarketParticipantAsync(chargesDatabaseWriteContext, MarketParticipantOwnerId);
+            var marketParticipantId = await QueryServiceAutoMoqDataFixer.GetOrAddMarketParticipantAsync(chargesDatabaseWriteContext);
 
             var charge = GetValidCharge(marketParticipantId);
 
@@ -296,7 +292,7 @@ namespace GreenEnergyHub.Charges.IntegrationTests.IntegrationTests.WebApi.QueryS
 
             await using var chargesDatabaseQueryContext = _databaseManager.CreateDbQueryContext();
             var expected = GetActiveCharges(chargesDatabaseQueryContext).Count(c => c.OwnerId == marketParticipantId);
-            var searchCriteria = new SearchCriteriaDtoBuilder()
+            var searchCriteria = new ChargeSearchCriteriaV1DtoBuilder()
                 .WithOwnerId(marketParticipantId)
                 .Build();
 
@@ -315,10 +311,10 @@ namespace GreenEnergyHub.Charges.IntegrationTests.IntegrationTests.WebApi.QueryS
         {
             // Arrange
             await using var chargesDatabaseWriteContext = _databaseManager.CreateDbContext();
-            var marketParticipantId = await GetOrAddMarketParticipantAsync(chargesDatabaseWriteContext, MarketParticipantOwnerId);
+            var marketParticipantId = await QueryServiceAutoMoqDataFixer.GetOrAddMarketParticipantAsync(chargesDatabaseWriteContext);
             var charge = GetValidCharge(marketParticipantId);
 
-            var marketParticipantId2 = await GetOrAddMarketParticipantAsync(chargesDatabaseWriteContext, "SecondMarketParticipantOwnerId");
+            var marketParticipantId2 = await QueryServiceAutoMoqDataFixer.GetOrAddMarketParticipantAsync(chargesDatabaseWriteContext, "SecondMarketParticipantOwnerId");
             var charge2 = GetValidCharge(marketParticipantId2);
 
             chargesDatabaseWriteContext.Charges.Add(charge);
@@ -327,7 +323,7 @@ namespace GreenEnergyHub.Charges.IntegrationTests.IntegrationTests.WebApi.QueryS
 
             await using var chargesDatabaseQueryContext = _databaseManager.CreateDbQueryContext();
             var expected = GetActiveCharges(chargesDatabaseQueryContext).Count(c => c.OwnerId == marketParticipantId || c.OwnerId == marketParticipantId2);
-            var searchCriteria = new SearchCriteriaDtoBuilder()
+            var searchCriteria = new ChargeSearchCriteriaV1DtoBuilder()
                 .WithOwnerIds(new List<Guid> { marketParticipantId, marketParticipantId2 })
                 .Build();
 
@@ -347,7 +343,7 @@ namespace GreenEnergyHub.Charges.IntegrationTests.IntegrationTests.WebApi.QueryS
         {
             // Arrange
             await using var chargesDatabaseWriteContext = _databaseManager.CreateDbContext();
-            var marketParticipantId = await GetOrAddMarketParticipantAsync(chargesDatabaseWriteContext, MarketParticipantOwnerId);
+            var marketParticipantId = await QueryServiceAutoMoqDataFixer.GetOrAddMarketParticipantAsync(chargesDatabaseWriteContext);
 
             var charge = GetValidCharge(marketParticipantId);
 
@@ -356,7 +352,7 @@ namespace GreenEnergyHub.Charges.IntegrationTests.IntegrationTests.WebApi.QueryS
 
             await using var chargesDatabaseQueryContext = _databaseManager.CreateDbQueryContext();
             var expected = GetActiveCharges(chargesDatabaseQueryContext).Count();
-            var searchCriteria = new SearchCriteriaDtoBuilder()
+            var searchCriteria = new ChargeSearchCriteriaV1DtoBuilder()
                 .WithOwnerIds(null!)
                 .Build();
 
@@ -375,7 +371,7 @@ namespace GreenEnergyHub.Charges.IntegrationTests.IntegrationTests.WebApi.QueryS
         {
             // Arrange
             await using var chargesDatabaseWriteContext = _databaseManager.CreateDbContext();
-            var marketParticipantId = await GetOrAddMarketParticipantAsync(chargesDatabaseWriteContext, MarketParticipantOwnerId);
+            var marketParticipantId = await QueryServiceAutoMoqDataFixer.GetOrAddMarketParticipantAsync(chargesDatabaseWriteContext);
 
             var charge = GetPlannedCharge(marketParticipantId);
             var todayAtMidnightUtc = DateTime.Now.Date.ToUniversalTime().ToInstant();
@@ -386,7 +382,7 @@ namespace GreenEnergyHub.Charges.IntegrationTests.IntegrationTests.WebApi.QueryS
                 c => c.Periods.Any(p => p.StartDateTime <= todayAtMidnightUtc));
 
             await using var chargesDatabaseQueryContext = _databaseManager.CreateDbQueryContext();
-            var searchCriteria = new SearchCriteriaDtoBuilder()
+            var searchCriteria = new ChargeSearchCriteriaV1DtoBuilder()
                 .Build();
 
             var sut = GetSut(chargesDatabaseQueryContext);
@@ -450,22 +446,6 @@ namespace GreenEnergyHub.Charges.IntegrationTests.IntegrationTests.WebApi.QueryS
                 .WithOwnerId(ownerId)
                 .AddPeriods(periods)
                 .Build();
-        }
-
-        private static async Task<Guid> GetOrAddMarketParticipantAsync(ChargesDatabaseContext context, string marketParticipantOwnerId)
-        {
-            var marketParticipant = await context
-                .MarketParticipants
-                .SingleOrDefaultAsync(x => x.MarketParticipantId == marketParticipantOwnerId);
-
-            if (marketParticipant != null)
-                return marketParticipant.Id;
-
-            marketParticipant = new TestGridAccessProvider(marketParticipantOwnerId);
-            context.MarketParticipants.Add(marketParticipant);
-            await context.SaveChangesAsync();
-
-            return marketParticipant.Id;
         }
 
         private static IQueryable<QueryApi.Model.Charge> GetActiveCharges(QueryDbContext chargesDatabaseQueryContext)